First pass on moving profiles to new app
