+++ oc_lettings_site/admin.py
@@ -2,9 +2,9 @@ from django.contrib import admin
 
 from .models import Letting
 from .models import Address
-from .models import Profile
+# from .models import Profile
 
 
 admin.site.register(Letting)
 admin.site.register(Address)
-admin.site.register(Profile)
+# admin.site.register(Profile)
