+++ oc_lettings_site/urls.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from django.urls import path
+from django.urls import path, include
 
 from . import views
 
@@ -7,7 +7,6 @@ urlpatterns = [
     path('', views.index, name='index'),
     path('lettings/', views.lettings_index, name='lettings_index'),
     path('lettings/<int:letting_id>/', views.letting, name='letting'),
-    path('profiles/', views.profiles_index, name='profiles_index'),
-    path('profiles/<str:username>/', views.profile, name='profile'),
+    path('profiles/', include('profiles.urls')),
     path('admin/', admin.site.urls),
 ]
