+++ profiles/admin.py
@@ -1,0 +1,6 @@
+from .models import Profile
+
+from django.contrib import admin
+
+
+admin.site.register(Profile)
