+++ templates/index.html
@@ -1,4 +1,4 @@
 <title>Holiday Homes</title>
 <h1>Welcome to Holiday Homes</h1>
-<div><a href="{% url 'profiles_index' %}">Profiles</a></div>
+<div><a href="{% url 'profiles:index' %}">Profiles</a></div>
 <div><a href="{% url 'lettings_index' %}">Lettings</a></div>
