+++ templates/profiles_index.html
@@ -4,7 +4,7 @@
     <ul>
     {% for profile in profiles_list %}
         <li>
-            <a href="{% url 'profile' username=profile.user.username %}">
+            <a href="{% url 'profiles:profile' username=profile.user.username %}">
                 {{ profile.user.username }}
             </a>
         </li>
